feat: replace mqsicreatebar with ibmint
Using the newer ACE images allows the mqsicreatebar command to be
replaced with the lighter-weight and faster ibmint command.

This removes the need for the image-builder pipeline as the
regular ACE image supports ibmint as-is, so this change makes the
overall pipeline project much simpler, and faster to run.

Huge thanks to @tdolby-at-uk-ibm-com for the prompt to make this
change.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>