docs: explain how to provide credentials
IBM Entitled Registry key

The pipeline needs a docker pull secret for pulling the ACE images
used for building and testing the bar files.

As people rarely read documentation, I wanted to make this really
obvious and self-documenting, so I've made the script check that
this has been provided and stop if it's missing.

Creating the Docker config json is a bit fiddly, so I think the
simplest way to do this is to leave it to the logic in kubectl's
"create secret docker-registry" method, so rather than ask for the
secret in a file, I'm doing it through an environment variable.

Github credentials

Not everyone will need this - the only reason why I started with
this is because I started working on the pipeline in github
enterprise.

So I made this file optional, modifying the script to apply it
only if it's provided.

I've updated the README to explain what is needed in the file for
people who do need the pipeline to have git credentials.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>